requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pytesseract==0.3.13
#tesseract da yüklü olmalı